+++ packages/aws-blocks/aws-blocks/index.handler.ts
@@ -1,3 +1,5 @@
 import { createLambdaHandler } from "@aws-blocks/blocks/lambda-handler";
 
-export const handler = createLambdaHandler(() => import("../src/index.js"));
+export const handler = createLambdaHandler(
+  () => import("../src/index.blocks.js")
+);